+++ javascript/node/selenium-webdriver/example/chrome_mobile_emulation.js
@@ -1,0 +1,37 @@
+// Copyright 2015 Selenium committers
+// Copyright 2015 Software Freedom Conservancy
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+//     You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+/**
+ * @fileoverview This is an example of emulating a mobile device using the
+ * ChromeDriver.
+ */
+
+var webdriver = require('..'),
+    By = webdriver.By,
+    until = webdriver.until,
+    chrome = require('../chrome');
+
+
+var driver = new webdriver.Builder()
+    .forBrowser('chrome')
+    .setChromeOptions(new chrome.Options()
+        .setMobileEmulation({deviceName: 'Google Nexus 5'}))
+    .build();
+
+driver.get('http://www.google.com/ncr');
+driver.findElement(By.name('q')).sendKeys('webdriver');
+driver.findElement(By.name('btnG')).click();
+driver.wait(until.titleIs('webdriver - Google Search'), 1000);
+driver.quit();